Refactor Kind.of?
