+++ lib/kind/modules.rb
@@ -1,10 +1,10 @@
 # frozen_string_literal: true
 
-require 'kind/modules/01_checker_utils'
-require 'kind/modules/02_checker_methods'
+require 'kind/modules/checker_utils'
+require 'kind/modules/checker_methods'
 
-require 'kind/modules/core/class'
 require 'kind/modules/core/array'
+require 'kind/modules/core/class'
 require 'kind/modules/core/comparable'
 require 'kind/modules/core/enumerable'
 require 'kind/modules/core/enumerator'
