+++ lib/kind/modules/checker_methods.rb
