+++ lib/kind/modules/checker_utils.rb
